moved DeleteRoomComplexAsync from RoomRepository to ComplexRepository; fixed references, injected IRoomRepository into ComplexRepository ctor
