+++ lodging/src/Revature.Lodging.Api/Controllers/RoomController.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using Revature.Lodging.Lib;
+using Revature.Lodging.Lib.Interface;
 using Revature.Lodging.Lib.Models;
 
 namespace Revature.Lodging.Api.Controllers
