+++ lodging/src/Revature.Lodging.Tests/DataTests/RepositoryTest.cs
@@ -142,9 +142,10 @@ namespace Revature.Lodging.Tests.DataTests
               .UseInMemoryDatabase("CreateComplexAsyncTest")
               .Options;
       using var testContext = new Entity.LodgingDbContext(options);
-      var repo = new ComplexRepository(testContext, log);
+      var roomRepo = new RoomRepository(testContext);
+      var complexRepo = new ComplexRepository(testContext, roomRepo, log);
 
-      var result = await repo.CreateComplexAsync(_complex1);
+      var result = await complexRepo.CreateComplexAsync(_complex1);
       var checker = testContext.Complex.First().Id;
 
       Assert.Equal(checker, CId1);
@@ -162,13 +163,14 @@ namespace Revature.Lodging.Tests.DataTests
               .UseInMemoryDatabase("ReadComplexListTest")
               .Options;
       using var testContext = new Entity.LodgingDbContext(options);
-      var repo = new ComplexRepository(testContext, log);
+      var roomRepo = new RoomRepository(testContext);
+      var complexRepo = new ComplexRepository(testContext, roomRepo, log);
 
       testContext.Add(_complexE1);
       testContext.Add(_complexE2);
       testContext.SaveChanges();
 
-      var list = await repo.ReadComplexListAsync();
+      var list = await complexRepo.ReadComplexListAsync();
 
       Assert.Equal("Liv+", list[0].ComplexName);
       Assert.Equal("9876543210", list[1].ContactNumber);
@@ -187,11 +189,12 @@ namespace Revature.Lodging.Tests.DataTests
               .UseInMemoryDatabase("ReadComplexAsyncTest")
               .Options;
       using var testContext = new Entity.LodgingDbContext(options);
-      var repo = new ComplexRepository(testContext, log);
+      var roomRepo = new RoomRepository(testContext);
+      var complexRepo = new ComplexRepository(testContext, roomRepo, log);
 
       testContext.Add(_complexE1);
 
-      var read = await repo.ReadComplexByIdAsync(CId1);
+      var read = await complexRepo.ReadComplexByIdAsync(CId1);
 
       Assert.Equal(CId1, read.Id);
     }
@@ -208,7 +211,8 @@ namespace Revature.Lodging.Tests.DataTests
               .UseInMemoryDatabase("UpdateComplexAsync")
               .Options;
       using var testContext = new Entity.LodgingDbContext(options);
-      var repo = new ComplexRepository(testContext, log);
+      var roomRepo = new RoomRepository(testContext);
+      var complexRepo = new ComplexRepository(testContext, roomRepo, log);
 
       testContext.Add(_complexE1);
       testContext.SaveChanges();
@@ -219,7 +223,7 @@ namespace Revature.Lodging.Tests.DataTests
         ComplexName = "Liv++",
         ContactNumber = "7894561231"
       };
-      _ = await repo.UpdateComplexAsync(update);
+      _ = await complexRepo.UpdateComplexAsync(update);
       var result = testContext.Complex.Find(CId1).ComplexName;
       var phone = testContext.Complex.Find(CId1).ContactNumber;
 
@@ -239,12 +243,13 @@ namespace Revature.Lodging.Tests.DataTests
               .UseInMemoryDatabase("DeleteComplexTest")
               .Options;
       using var testContext = new Entity.LodgingDbContext(options);
-      var repo = new ComplexRepository(testContext, log);
+      var roomRepo = new RoomRepository(testContext);
+      var complexRepo = new ComplexRepository(testContext, roomRepo, log);
 
       testContext.Add(_complexE1);
       testContext.Add(_complexE2);
 
-      var status = await repo.DeleteComplexAsync(CId1);
+      var status = await complexRepo.DeleteComplexAsync(CId1);
 
       var result = testContext.Complex.First().ComplexName;
       var phone = testContext.Complex.First().ContactNumber;
@@ -405,7 +410,8 @@ namespace Revature.Lodging.Tests.DataTests
               .UseInMemoryDatabase("ReadComplexByProviderIDTest")
               .Options;
       using var testContext = new Entity.LodgingDbContext(options);
-      var repo = new ComplexRepository(testContext, log);
+      var roomRepo = new RoomRepository(testContext);
+      var complexRepo = new ComplexRepository(testContext, roomRepo, log);
 
       var complexE3 = new Entity.Complex
       {
@@ -421,7 +427,7 @@ namespace Revature.Lodging.Tests.DataTests
       testContext.Add(complexE3);
       testContext.SaveChanges();
 
-      var complices = await repo.ReadComplexByProviderIdAsync(PId1);
+      var complices = await complexRepo.ReadComplexByProviderIdAsync(PId1);
 
       Assert.Equal("Liv+", complices[0].ComplexName);
       Assert.Equal("XXX", complices[1].ComplexName);
@@ -507,7 +513,8 @@ namespace Revature.Lodging.Tests.DataTests
               .UseInMemoryDatabase("DeleteAmenityAsyncTest")
               .Options;
       using var testContext = new Entity.LodgingDbContext(options);
-      var repo = new ComplexRepository(testContext, log);
+      var roomRepo = new RoomRepository(testContext);
+      var complexRepo = new ComplexRepository(testContext, roomRepo, log);
 
       var name = "Liv+";
       var phone = "1234567890";
@@ -515,7 +522,7 @@ namespace Revature.Lodging.Tests.DataTests
       testContext.Add(_complexE1);
       testContext.SaveChanges();
 
-      var check = await repo.ReadComplexByNameAndNumberAsync(name, phone);
+      var check = await complexRepo.ReadComplexByNameAndNumberAsync(name, phone);
 
       Assert.Equal(name, check.ComplexName);
       Assert.Equal(phone, check.ContactNumber);
